Remove the class for kernel arguments from pyanaconda.flags
Move the BootArgs class to pyanaconda.core.kernel and rename it to
KernelArguments. The __init__ method is simplified and the logic
is replaced with methods from_defaults and from_string.
