+++ pyanaconda/flags.py
@@ -18,14 +18,12 @@
 #
 
 import selinux
-import shlex
-import glob
-from pyanaconda.core.constants import SELINUX_DEFAULT, CMDLINE_APPEND, CMDLINE_LIST, \
-    ANACONDA_ENVIRON
-from collections import OrderedDict
+from pyanaconda.core.constants import SELINUX_DEFAULT, ANACONDA_ENVIRON
+from pyanaconda.core.kernel import KernelArguments
 
 from pyanaconda.anaconda_loggers import get_module_logger
 log = get_module_logger(__name__)
+
 
 # A lot of effort, but it only allows a limited set of flags to be referenced
 class Flags(object):
@@ -83,7 +81,7 @@ class Flags(object):
         # current runtime environments
         self.environs = [ANACONDA_ENVIRON]
         # parse the boot commandline
-        self.cmdline = BootArgs()
+        self.cmdline = KernelArguments.from_defaults()
         # Lock it down: no more creating new flags!
         self.__dict__['_in_init'] = False
         if read_cmdline:
@@ -97,108 +95,6 @@ class Flags(object):
         if not selinux.is_selinux_enabled():
             self.selinux = 0
 
-cmdline_files = ['/proc/cmdline', '/run/install/cmdline',
-                 '/run/install/cmdline.d/*.conf', '/etc/cmdline']
-class BootArgs(OrderedDict):
-    """
-    Hold boot arguments as an OrderedDict.
-    """
-    def __init__(self, cmdline=None, files=None):
-        """
-        Create a BootArgs object.
-        Reads each of the "files", then parses "cmdline" if it was provided.
-        """
-        super().__init__()
-        if files is None:
-            self.read(cmdline_files)
-        elif files:
-            self.read(files)
-        if cmdline:
-            self.readstr(cmdline)
-
-    def read(self, filenames):
-        """
-        Read and parse a filename (or a list of filenames).
-        Files that can't be read are silently ignored.
-        Returns a list of successfully read files.
-        filenames can contain \\*, ?, and character ranges expressed with []
-        """
-
-        readfiles = []
-        if isinstance(filenames, str):
-            filenames = [filenames]
-
-        # Expand any filename globs
-        filenames = [f for g in filenames for f in glob.glob(g)]
-
-        for f in filenames:
-            try:
-                self.readstr(open(f).read())
-                readfiles.append(f)
-            except IOError:
-                continue
-        return readfiles
-
-    def readstr(self, cmdline):
-        cmdline = cmdline.strip()
-        # if the BOOT_IMAGE contains a space, pxelinux will strip one of the
-        # quotes leaving one at the end that shlex doesn't know what to do
-        # with
-        (left, middle, right) = cmdline.rpartition("BOOT_IMAGE=")
-        if right.count('"') % 2:
-            cmdline = left + middle + '"' + right
-
-        # shlex doesn't properly handle \\ (it removes them)
-        # which scrambles the spaces used in labels so use underscores
-        cmdline = cmdline.replace("\\x20", "_")
-
-        lst = shlex.split(cmdline)
-
-        # options might have the inst. prefix (used to differentiate
-        # boot options for the installer from other boot options)
-        inst_prefix = "inst."
-
-        for i in lst:
-            # drop the inst. prefix (if found), so that getbool() works
-            # consistently for both "foo=0" and "inst.foo=0"
-            if i.startswith(inst_prefix):
-                i = i[len(inst_prefix):]
-
-            if "=" in i:
-                (key, val) = i.split("=", 1)
-            else:
-                key = i
-                val = None
-
-            # Some duplicate args create a space separated string
-            if key in CMDLINE_APPEND and self.get(key, None):
-                if val:
-                    self[key] = self[key] + " " + val
-            # Some arguments can contain spaces so adding them in one string is not that helpful
-            elif key in CMDLINE_LIST:
-                if val:
-                    if not self.get(key, None):
-                        self[key] = []
-                    self[key].append(val)
-            else:
-                self[key] = val
-
-    def getbool(self, arg, default=False):
-        """
-        Return the value of the given arg, as a boolean. The rules are:
-        - "arg", "arg=val": True
-        - "noarg", "noarg=val", "arg=[0|off|no]": False
-        """
-        result = default
-        for a in self:
-            if a == arg:
-                if self[arg] in ("0", "off", "no"):
-                    result = False
-                else:
-                    result = True
-            elif a == 'no' + arg:
-                result = False  # XXX: should noarg=off -> True?
-        return result
 
 def can_touch_runtime_system(msg, touch_live=False):
     """
@@ -225,5 +121,6 @@ def can_touch_runtime_system(msg, touch_live=False):
         return False
 
     return True
+
 
 flags = Flags()
